PR-E-8: extract TypeDefEmitter (TypeDef shape + constructors)
Moves the 12 TypeDef-shape and constructor emission methods (plus 5
cluster-exclusive helpers) out of ReflectionMetadataEmitter into a new
internal sealed TypeDefEmitter at
src/Core/CodeAnalysis/Emit/TypeDefEmitter.cs, continuing the Phase-2
ReflectionMetadataEmitter decomposition.

TypeDef shape methods moved (former line numbers):
  EmitStructTypeDef         (~2020)
  EmitNestedStructTypeDef   (~2482)
  EmitEnumTypeDef           (~2587)
  EmitInterfaceTypeDef      (~3503)
  EmitDelegateTypeDef       (~3541)
  EmitAbstractMethod        (~3687)

Constructor methods moved (former line numbers):
  EmitClassDefaultConstructor             (~3720)
  EmitStaticConstructor                   (~3753)
  EmitClassPrimaryConstructor             (~3942)
  EmitClassConstructorWithBaseInitializer (~4013)
  EmitClassConstructorWithBody            (~4183)
  EmitDefaultConstructor                  (~4871)

Private helpers moved (cluster-exclusive):
  GetBaseCtorToken
  GetBaseInitializerCtorToken
  GetImportedBaseDefaultCtorReference
  EmitIsReadOnlyAttribute
  EmitIsByRefLikeAttribute

What stays on ReflectionMetadataEmitter (deferred to later PRs):
  - EmitBaseConstructorArguments is a method of the nested BodyEmitter
    class; it moves with BodyEmitter in PR-E-11.
  - EmitInlineStructConstructor already moved to DataStructSynthesizer
    in PR-E-6.
  - Emit / EmitCore / EmitGlobalFieldDefs / EmitFunction and the
    assembly-level Emit*Attribute family stay on the thinned root for
    PR-E-12.
  - BodyEmitter and closure/state-machine emission are PR-E-9 / E-10 /
    E-11.

Wiring:
  - ReflectionMetadataEmitter gets a private TypeDefEmitter typeDefEmitter
    field instantiated in EmitCore right after memberDefEmitter.
  - All 17 internal call sites are rewritten to this.typeDefEmitter.X(...).
  - TypeDefEmitter depends on EmitContext, MetadataTokenCache,
    WellKnownReferences, and is composed with seven Action/Func callbacks
    onto the still-private helpers on the root (EncodeTypeSymbol,
    EncodeReturnSymbol, GetTypeReference, NextParameterHandle,
    EmitUserAttributes, EmitIsReadOnlyAttributeOnParameter,
    GetCtorReference).
  - Three BodyEmitter-driven constructor bodies are emitted via three new
    *BodyBytes helpers on the root (kept adjacent to the BodyEmitter nested
    class) and passed as Func<...> callbacks; this avoids widening
    BodyEmitter's visibility until PR-E-11.
  - MapTypeAccessibility / MapFieldAccessibility are duplicated as private
    statics in TypeDefEmitter (root retains its own copies for
    EmitGlobalFieldDefs / EmitFunction) — two trivial switch expressions
    is cheaper than widening visibility per PR-0's risk note.

Validation:
  - dotnet build GSharp.sln: green.
  - Core.Tests: 2120 passed.
  - RefactoringBaselineTests: zero hash diffs.
  - DeterministicEmitTests: green.
  - Interpreter.Tests: 72, Sdk.Tests: 24, LanguageServer.Tests: 242,
    Compiler.Tests (all classes individually): 0 failures.
  - No public or internal API surface visible to Compilation.cs or any
    test has been renamed.

LoC delta:
  - ReflectionMetadataEmitter.cs: 14229 -> 13249 (-980)
  - TypeDefEmitter.cs (new): 1256

Next: PR-E-9 ClosureEmitter.
